requirements.txt
flask
python-dotenv
# opencv-python-headless
numpy
tensorflow-cpu==2.17.0
scikit-learn
h5py
pandas
celery[redis]
redis
tf2onnx
onnxruntime==1.17.0
pillow
pytesseract
opencv-python
argostranslate
tqdm
python-docx
scikit-image
python-magic 